requirements.txt
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.5.1
hf-xet==1.1.2
huggingface-hub==0.32.2
idna==3.10
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
protobuf==6.31.0
PyMuPDF==1.26.0
pyTelegramBotAPI==4.27.0
python-docx==1.1.2
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==80.9.0
SQLAlchemy==2.0.41
sympy==1.14.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typing_extensions==4.13.2
urllib3==2.4.0
Whoosh==2.7.4
